ref: Move conversion func to plugin.py
